Formatting - ICO updated from string to embeds
