+++ commands/personal.ts
@@ -1,6 +1,6 @@
 import { User } from '../db/models/user'
 import { lookupCoin } from './cmc/api'
-import { portfolioEmbed, createPortfolioMessage } from './cmc/embedders'
+import { portfolioEmbed, createPortfolioMessage } from './cmc/formatters'
 const addCoin = async ([coinName, unitsOwned, username]) => {
 	const { id: coinId } = lookupCoin(coinName)
 	if (!coinId) {
